# openapi/personamy-api.yaml
openapi: 3.0.3
info:
  title: Adele ai Backend API
  description: |
    API complète pour l'analyse de données financières et d'options.
    
    Architecture modulaire en 3 couches :
    - **Ingestion** : Collecte rapide et idempotente des données externes
    - **Analyse unitaire** : Analyses par module (règles simples + IA)
    - **Synthèse globale** : Analyses combinées et insights avancés
  version: 1.0.0
  contact:
    name: Adele ai API Support
servers:
  - url: https://tsdd1sibd1.execute-api.eu-west-3.amazonaws.com/prod
    description: API Gateway 1 - Application Principale (`@baseUrlMain` dans `api-tests.http`)
  - url: https://faq9dl95v7.execute-api.eu-west-3.amazonaws.com/prod
    description: API Gateway 2 - Données Brutes FMP + Unusual Whales (`@baseUrlData` dans `api-tests.http`)
...
security:
  - bearerAuth: []
paths:
...
  /ingest/flow-alerts:
    post:
      tags: [Ingestion]
      summary: Ingérer les flow alerts dans la table flow_alerts (pour analyse pro)
      description: |
        **NOUVEAU** : Route d'ingestion dédiée pour l'analyse pro du flow options.
        
        Appelle l'API Unusual Whales `/option-trades/flow-alerts`, normalise et stocke les données dans la table `flow_alerts`.
        Cette route est **indispensable avant** d'utiliser `/ai/flow-options-analysis-pro` pour éviter les cold starts et latences.
        
        **Différence avec `/ingest/options-flow-alerts`** :
        - Cette route stocke dans `flow_alerts` (table dédiée pour l'analyse pro)
        - `/ingest/options-flow-alerts` stocke dans `options_flow` avec un flag
        
        **Workflow recommandé** :
        1. `POST /ingest/flow-alerts?ticker=MSFT` (ingestion)
        2. `POST /ai/flow-options-analysis-pro` avec `{"ticker": "MSFT"}` (analyse depuis DB)
      parameters:
        - name: ticker
          in: query
          required: true
          schema:
            type: string
            example: MSFT
          description: Symbole du ticker
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: "Nombre maximum de flow alerts à récupérer (max: 200)"
        - name: min_premium
          in: query
          required: false
          schema:
            type: number
            format: float
            default: 10000
          description: "Prime minimum en USD (défaut: 10k$)"
        - name: is_call
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Inclure les calls
        - name: is_put
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Inclure les puts
        - name: is_sweep
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Inclure les sweeps
        - name: is_floor
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Inclure les floor trades
      responses:
        '200':
          description: Flow alerts ingérées avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  ticker:
                    type: string
                    example: MSFT
                  module:
                    type: string
                    example: flow_alerts
                  count:
                    type: integer
                    example: 15
                    description: Nombre de flow alerts ingérées
                  state:
                    $ref: '#/components/schemas/ModuleState'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Paramètre manquant ou invalide
        '500':
          description: Erreur serveur
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Cognito
  schemas:
...
    ModuleState:
      type: object
      properties:
        ticker:
          type: string
          example: NVDA
        module_id:
          type: string
          example: options_flow
        status:
          type: string
          enum: [missing, refreshing, ready, error]
        fetched_at:
          type: string
          format: date-time
        data_date:
          type: string
          format: date
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          properties:
            count:
              type: integer
            call_count:
              type: integer
            put_count:
              type: integer
            ingested_at:
              type: string
              format: date-time
        refresh_lock_until:
          type: string
          format: date-time
          nullable: true
        refresh_lock_owner:
          type: string
          nullable: true
...